requirements.txt
django<1.3
django-debug-toolbar
django-timezones
django-gravatar
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django-registration
pytz
pil
python-dateutil
django-extensions

